directory redesign and css refactoring
