+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import LandingPage from "./landingPage/landingPage";
-import GamePage from "./gamePage/gamePage";
+import LandingPage from "./pages/landingPage";
+import GamePage from "./pages/gamePage";
 import "./App.css";
 
 function App() {
