+++ src/components/gameResultModal.js
@@ -1,7 +1,5 @@
 import React from "react";
-import "./gameResultModal.css";
-import LoseGame from "../gameResult/loseGame";
-import WonGame from "../gameResult/wonGame";
+import "./styles.css";
 
 function GameResultModal({ gameState }) {
   const { userLost, userFinished, revealedWord } = gameState;
@@ -10,11 +8,15 @@ function GameResultModal({ gameState }) {
     window.location.reload(false);
   }
 
+  const WonGame = () => <div className="wonGame">YOU WIN!</div>;
+
+  const LoseGame = () => <div className="lostGame">YOU LOST!</div>;
+
   return (
     userFinished && (
       <div className="gameResultModal">
         <h2>{userLost ? <LoseGame /> : <WonGame />}</h2>
-        <p className="line">The word was:</p>
+        <p className="wordWas">The word was:</p>
         <p className="word">{revealedWord}</p>
         <input
           autoFocus
