+++ src/components/logoSpin.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import logo from './logo.svg';
+import logo from '../assets/logo.svg';
 import '../App.css';
