--- src/gamePage/guessedLetters.css
@@ -1,26 +1,0 @@
-.guessedLetters {
-    display: flex;
-    justify-content: center;
-    gap: 0.25em;
-    font-size: 2rem;
-    font-style: bold;
-    text-transform: uppercase;
-    font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-      Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-    height: 30px;
-  }
-
-  .guessedLettersHeader {
-    padding-top: 20px;
-    color: aqua;
-    font-size: 1rem;
-    font-style: bold;
-    text-transform: uppercase;
-    font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-      Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  }
-
-.guessesRemainingProps {
-  color: red;
-  font-weight: bold;
-}
