--- src/gamePage/remainingGuesses.css
@@ -1,8 +1,0 @@
-.guessesRemaining {
-    font-size: 13px;
-    font-style: italic;
-    padding: 4px;
-    color: beige;
-  }
-  
-
