--- src/gamePage/revealedLetters.css
@@ -1,24 +1,0 @@
-.letters {
-  display: flex;
-  gap: 0.25em;
-  font-size: 6rem;
-  font-weight: bold;
-  text-transform: uppercase;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-    Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  padding: 20px;
-}
-
-.spaces {
-  border-bottom: 0.1em solid beige;
-  border-radius: 1px;
-}
-
-
-.hidden {
-  visibility: hidden;
-}
-
-.visible {
-  visibility: visible;
-}
