--- src/gameResult/gameResultModal.css
@@ -1,26 +1,0 @@
-.gameResultModal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.819);
-    backdrop-filter: blur(5px);
-    z-index: 100;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.line {
-    font-size: 20px;
-    font-weight: bold;
-}
-
-.word {
-    font-size: 2rem;
-    font-weight: bold;
-    color: rgb(42, 200, 224);  
-    text-transform: uppercase;
-}
