--- src/gameResult/loseGame.js
@@ -1,20 +1,0 @@
-
-function LoseGame() {
-
-    return (
-        <div style={{
-            color: "red",
-            fontSize: "65px",
-            fontWeight: "bold"
-        }}>
-            YOU LOST!
-        </div>
-    )
-
-}
-
-
-
-
-
-export default LoseGame;
