--- src/gameResult/wonGame.js
@@ -1,22 +1,0 @@
-
-function WonGame() {
-
-    return (
-        <div
-        style={{
-            color: "green",
-            fontSize: "65px",
-            fontWeight: "bold"
-        }}
-        >
-            YOU WIN!
-        </div>
-    )
-
-}
-
-
-
-
-
-export default WonGame;
