--- src/guessrForm/guessrForm.css
@@ -1,61 +1,0 @@
-.form {
-  padding-top: 5px;
-}
-
-.guessTitle {
-  font-size: 18px;
-}
-
-.inputBox {
-  padding-bottom: 5px;
-}
-
-.inputField {
-  margin: 8px;
-  width: 32px;
-  font-size: 1.7rem;
-  border-radius: 6px;
-  border-style: solid;
-  text-align: center;
-  resize: none;
-  background-color: beige;
-  color: #282c34;
-}
-
-textarea:focus,
-input:focus,
-input[type]:focus,
-.uneditable-input:focus {
-  border-color: rgb(42, 200, 224);
-  box-shadow: 0 1px 10px rgba(27, 109, 116, 0.075) inset,
-    0 0 8px rgba(24, 149, 156, 0.6);
-  outline: 0 none;
-}
-
-.remainingGuesses {
-  font-size: 14px;
-  font-style: italic;
-  padding: 4px;
-  color: beige;
-}
-
-.submitButton {
-  color: beige;
-  background-color: #282c34;
-  border-style: solid;
-  font-size: medium;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.playAgainButton {
-  color: rgb(42, 200, 224);
-  background-color: #282c34;
-  border-style: solid;
-  border-color: rgb(42, 200, 224);
-  box-shadow: 0 10px 10px rgba(27, 109, 116, 0.075) inset,
-    0 0 8px rgba(24, 149, 156, 0.6);
-  font-size: medium;
-  border-radius: 3px;
-  cursor: pointer;
-}
