--- src/landingPage/landingPage.css
@@ -1,8 +1,0 @@
-.wrapper {
-    display: flex;
-    width: 100vw;
-    min-height: 100vh;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
