+++ src/pages/styles.scss
@@ -1,12 +1,20 @@
-.wrapper {
+.pageWrapper {
     display: flex;
     width: 100vw;
     min-height: 100vh;
     justify-content: center;
     align-items: center;
     flex-direction: column;
+  }
+  
+  .gamePageWrapper {
+    @extend .pageWrapper;
     color: beige;
     font-size: larger;
-    padding-bottom: 40px;
-    padding-top: 20%;
-}
+    padding: 20% 0 40px;
+  }
+  
+  .landingPageWrapper {
+    @extend .pageWrapper;
+  }
+
